+++ app/src/main/java/eu/kanade/tachiyomi/ui/player/mpv/MPVModels.kt
@@ -88,6 +88,8 @@ data class TrackNode(
 
     fun getMetadata(key: String): String? = metadata?.get(key)
     fun hasMetadata(): Boolean = !metadata.isNullOrEmpty()
+
+    fun getLanguage(): String? = (lang ?: metadata?.get("language"))?.takeUnless { it == "und" }
 }
 
 @Serializable
@@ -115,6 +117,7 @@ sealed interface VideoTrack {
         val data: Track,
         val index: Int,
         val id: Int? = null,
+        val language: String? = null,
         val mainSelection: Int = -1,
         val state: TrackState = TrackState.Idle,
     ) : VideoTrack
@@ -127,7 +130,7 @@ sealed interface VideoTrack {
 
     val lang: String
         get() = when (this) {
-            is External -> data.lang
+            is External -> language.orEmpty()
             is Internal -> data.lang.orEmpty()
         }